Set up properly credentials to convert using k525.
The k525 protocol and server distinguishes three parties - the originator,
source and target. The originator and source are almost the same for the PBS
use-case (since the PBS service key is used for both) but it's necessary that
the source server principal is from the final realm so we need to get another
ticket before the conversion takes place.