Add support for multi-realm setup to the server.
The daemon is now able to read multiple Kerberos databases, which are listed in
kdc.conf. The actual database is selected based on the realm of the target
principal from the converstion request.

The change renders quite useles the krb525 stanza of kdc.conf (pointing to the
database and key-file) and therefore its use is dropped altogether and only the
standard KDC locations are read.